Dynamically handle the URI
Generate the EU-SOL URI using the EU-SOL identifier provided by the user

Geneate the CGN URI by retrieving the CGN identifier in the information collected on
the EU-SOL page.

This remove the data that was hard-coded until now.

Note: We use at the moment two URL for EU-SOL, one as URL to collect the information
the other as URI when parsing the information retrieved.